+++ stratosphere/src/services/aws/customerprofiles.rs
@@ -52,7 +52,7 @@ pub mod calculatedattributedefinition {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-customerprofiles-calculatedattributedefinition-conditions.html
     pub struct Conditions_ {
-        pub object_count: Option<i64>,
+        pub object_count: Option<i32>,
         pub range: Option<Box<Range_>>,
         pub threshold: Option<Box<Threshold_>>,
     }
@@ -91,7 +91,7 @@ pub mod calculatedattributedefinition {
         pub timestamp_format: Option<crate::value::ExpString>,
         pub timestamp_source: Option<crate::value::ExpString>,
         pub unit: crate::value::ExpString,
-        pub value: Option<i64>,
+        pub value: Option<i32>,
         pub value_range: Option<Box<ValueRange_>>,
     }
     #[doc(hidden)]
@@ -137,7 +137,7 @@ pub mod calculatedattributedefinition {
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-customerprofiles-calculatedattributedefinition-readiness.html
     pub struct Readiness_ {
         pub message: Option<crate::value::ExpString>,
-        pub progress_percentage: Option<i64>,
+        pub progress_percentage: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -196,8 +196,8 @@ pub mod calculatedattributedefinition {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-customerprofiles-calculatedattributedefinition-valuerange.html
     pub struct ValueRange_ {
-        pub end: i64,
-        pub start: i64,
+        pub end: i32,
+        pub start: i32,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -538,8 +538,8 @@ pub mod domain {
         pub enabled: crate::value::ExpBool,
         pub exporting_config: Option<Box<ExportingConfig_>>,
         pub matching_rules: Option<Vec<MatchingRule_>>,
-        pub max_allowed_rule_level_for_matching: Option<i64>,
-        pub max_allowed_rule_level_for_merging: Option<i64>,
+        pub max_allowed_rule_level_for_matching: Option<i32>,
+        pub max_allowed_rule_level_for_merging: Option<i32>,
         pub status: Option<crate::value::ExpString>,
     }
     #[doc(hidden)]
@@ -715,7 +715,7 @@ pub mod eventtrigger {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-customerprofiles-eventtrigger-eventtriggerlimits.html
     pub struct EventTriggerLimits_ {
-        pub event_expiration: Option<i64>,
+        pub event_expiration: Option<i32>,
         pub periods: Option<Vec<Period_>>,
     }
     #[doc(hidden)]
@@ -786,10 +786,10 @@ pub mod eventtrigger {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-customerprofiles-eventtrigger-period.html
     pub struct Period_ {
-        pub max_invocations_per_profile: Option<i64>,
+        pub max_invocations_per_profile: Option<i32>,
         pub unit: crate::value::ExpString,
         pub unlimited: Option<crate::value::ExpBool>,
-        pub value: i64,
+        pub value: i32,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -1077,7 +1077,7 @@ pub mod integration {
         pub first_execution_from: Option<f64>,
         pub schedule_end_time: Option<f64>,
         pub schedule_expression: crate::value::ExpString,
-        pub schedule_offset: Option<i64>,
+        pub schedule_offset: Option<i32>,
         pub schedule_start_time: Option<f64>,
         pub timezone: Option<crate::value::ExpString>,
     }
@@ -2031,8 +2031,8 @@ pub mod segmentdefinition {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-customerprofiles-segmentdefinition-rangeoverride.html
     pub struct RangeOverride_ {
-        pub end: Option<i64>,
-        pub start: i64,
+        pub end: Option<i32>,
+        pub start: i32,
         pub unit: crate::value::ExpString,
     }
     #[doc(hidden)]
@@ -2201,7 +2201,7 @@ impl crate::template::ToResource for CalculatedAttributeDefinition_ {
 pub struct Domain_ {
     pub dead_letter_queue_url: Option<crate::value::ExpString>,
     pub default_encryption_key: Option<crate::value::ExpString>,
-    pub default_expiration_days: i64,
+    pub default_expiration_days: i32,
     pub domain_name: crate::value::ExpString,
     pub matching: Option<super::customerprofiles::domain::Matching_>,
     pub rule_based_matching: Option<super::customerprofiles::domain::RuleBasedMatching_>,
@@ -2455,10 +2455,10 @@ pub struct ObjectType_ {
     pub description: crate::value::ExpString,
     pub domain_name: crate::value::ExpString,
     pub encryption_key: Option<crate::value::ExpString>,
-    pub expiration_days: Option<i64>,
+    pub expiration_days: Option<i32>,
     pub fields: Option<Vec<super::customerprofiles::objecttype::FieldMap_>>,
     pub keys: Option<Vec<super::customerprofiles::objecttype::KeyMap_>>,
-    pub max_profile_object_count: Option<i64>,
+    pub max_profile_object_count: Option<i32>,
     pub object_type_name: crate::value::ExpString,
     pub source_last_updated_timestamp_format: Option<crate::value::ExpString>,
     pub tags: Option<Vec<crate::Tag_>>,
